Fix: [Server][LiveLLHLSSegmenter] H.265 で再生できない不具合を修正 target_duration を GOP 長に合わせる必要があったらしい？

+++ server/app/utils/hls/m3u8.py
@@ -17,9 +17,9 @@ from app.utils.hls.segment import Segment
 
 class M3U8:
 
-    def __init__(self, target_duration: int, part_target: float, list_size: int | None, hasInit: bool = False, prefix: str = '') -> None:
+    def __init__(self, target_duration: float, part_target: float, list_size: int | None, hasInit: bool = False, prefix: str = '') -> None:
         self.media_sequence: int = 0
-        self.target_duration: int = target_duration
+        self.target_duration: float = target_duration
         self.part_target: float = part_target
         self.list_size: int | None = list_size
         self.hasInit: bool = hasInit
